Refactored startdatelabel
